Fix some of the French emoji names not being matched
Some of the characters used in the French emoji names are currently not included in the regular expression for matching emoji names. One example is œ.

As word characters aren't continuous in Unicode it is impossible to correctly match them via a range expression. Thus, I converted the existing range to the word characters group which should be future proof regarding additional languages and avoids accidentally including non-word characters that could have been included by an overly vast range.