Add support for Multi-person skintones #204
The logic from demojize() is moved to two separate function tokenize and filter_tokens in a new file emoji/tokenizer.py
Also the logic for the search tree is moved to that file.

A new public function analyze() is available, that supports the multi-person skintones

The handling of the multi-person skintones can be controlled by the new `emoji.config` class, which is a static class that works as a module-wide configuration.